Add comments and improve code readability
